+++ db/migrate/20170501210143_add_time_of_day_format_to_variables.rb
@@ -1,0 +1,5 @@
+class AddTimeOfDayFormatToVariables < ActiveRecord::Migration[5.0]
+  def change
+    add_column :variables, :time_of_day_format, :string, null: false, default: '24hour'
+  end
+end